requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7       
charset-normalizer==3.0.1
colorama==0.4.6
Django==4.1.5
django-debug-toolbar==3.8.1
django-environ==0.9.0
django-factory-boy==1.0.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.1
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.4.1
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pytest==7.2.1
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
zipp==3.11.0